Warn about unknown placeholders directly after submitting a job
